// Stage Drive/Dashboard/justbitteam-stagedriving_dash-4cdbe67ac041/src/app/api/model/eventDTO.ts
/**
 * Stage Driving API
 * API specifications
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ActionDTO } from './actionDTO';
import { BookingDTO } from './bookingDTO';
import { CheckinDTO } from './checkinDTO';
import { CoordinateDTO } from './coordinateDTO';
import { EventInterestDTO } from './eventInterestDTO';
import { ImageDTO } from './imageDTO';


/**
 * Event model
 */
export interface EventDTO {
    id?: string;
    created?: string;
    modified?: string;
    name: string;
    description: string;
    coordinate: CoordinateDTO;
    address: string;
    country: string;
    city: string;
    town: string;
    zipcode: string;
    start: string;
    finish: string;
    organizer: string;
    distance?: number;
    parking?: string;
    website?: string;
    capacity?: number;
    status?: string;
    category?: string;
    visible?: boolean;
    code?: string;
    level?: string;
    mobile?: string;
    phone?: string;
    ncheckin?: number;
    nlike?: number;
    nride?: number;
    ncomment?: number;
    nbooking?: number;
    nbookingticket?: number;
    checkins?: Array<CheckinDTO>;
    bookings?: Array<BookingDTO>;
    images?: Array<ImageDTO>;
    actions?: Array<ActionDTO>;
    likes?: Array<ActionDTO>;
    comments?: Array<ActionDTO>;
    interests?: Array<EventInterestDTO>;
}