// Stage Drive/Dashboard/justbitteam-stagedriving_dash-4cdbe67ac041/src/app/api/model/stgdrvResponseDTO.ts
/**
 * Stage Driving API
 * API specifications
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface StgdrvResponseDTO {
    id?: string;
    created?: string;
    modified?: string;
    code?: number;
    message?: string;
    uri?: string;
}