// Stage Drive/Dashboard/justbitteam-stagedriving_dash-4cdbe67ac041/src/app/api/model/objectDTO.ts
/**
 * Stage Driving API
 * API specifications
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface ObjectDTO {
    id?: string;
    created?: string;
    modified?: string;
    content?: string;
    bucket?: string;
    mimeType?: string;
    size?: number;
}